before sell mac

+++ together/Event.swift
@@ -17,11 +17,11 @@ class Event {
     var photo: UIImage
     var category: String
     var ownerId: UInt64
-    var likes: Array<Int>
+    var likes: Array<UInt64>
     var location: String
     var startTime: Date
     var endTime: Date
-    var signedUsers: Array<Int>
+    var signedUsers: Array<UInt64>
     var linkUrls: Array<String>
     var linkStrings: Array<String>
     
@@ -34,10 +34,10 @@ class Event {
          contrebuted: Int = 0,
          category: String = "",
          ownerId: UInt64 = 0,
-         likes: Array<Int> = [0],
+         likes: Array<UInt64> = [0],
          location: String = "", startTime: Date = Date(),
          endTime: Date = Date(),
-         signedUsers: Array<Int> = [0],
+         signedUsers: Array<UInt64> = [0],
          linkUrls: Array<String> = [""],
          linkStrings: Array<String> = [""]
         )
